parser-json-simple: separate module for SimpleTreeDecoder
